removed antistuck from dismount
As antistuck is currently, it is expecting a target and when dismounting there may not always be a target selected. I added a timeout check which should be fine, navmesh has already been stopped and the only thing the dismount function is doing is trying to dismount so i think it should be safe if it took more than 10 seconds to dismount to try and teleport to reset itself.